backend/requirements.txt
asgiref==3.3.4
dataclasses==0.8
Django==3.0.5
djangorestframework==3.12.4
djongo==1.3.4
dnspython==2.1.0
importlib-metadata==4.0.1
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
pymongo==3.11.4
pytz==2021.1
sqlparse==0.2.4
typing-extensions==3.10.0.0
zipp==3.4.1
